Refactor of matching
