+++ wooting-macro-backend/src/grabbing/mod.rs
@@ -1,5 +1,4 @@
 pub mod executor;
 pub mod linux;
 pub mod macos;
-pub mod matcher;
 pub mod windows;
