+++ wooting-macro-backend/src/lib.rs
@@ -15,11 +15,12 @@ use crate::config::{ApplicationConfig, ConfigFile};
 use crate::config::CONFIG_DIR;
 use crate::grabbing::executor::input::macro_executor;
 #[cfg(target_os = "linux")]
-use crate::grabbing::linux::input;
+use crate::grabbing::linux::input::input;
 #[cfg(target_os = "macos")]
-use crate::grabbing::macos::input;
+use crate::grabbing::macos::input::input;
 #[cfg(target_os = "windows")]
-use crate::grabbing::windows::input;
+use crate::grabbing::windows::input::input;
+
 use crate::hid_table::*;
 use crate::macros::events::triggers::{MacroIndividualCommand, MacroTriggerEvent};
 use crate::macros::macro_data::{MacroData, MacroLookup};
