+++ wooting-macro-backend/src/macros/events/actions.rs
@@ -2,7 +2,6 @@ use crate::hid_table::HID_TO_RDEV;
 use crate::plugin::delay::DEFAULT_DELAY;
 use crate::plugin::{delay, key_press, mouse, phillips_hue, system_event};
 
-
 use std::time;
 use tokio::sync::mpsc::UnboundedSender;
