+++ wooting-macro-backend/src/macros/macros.rs
@@ -1,4 +1,3 @@
-
 use log::*;
 use rdev::EventType;
 use tokio::sync::mpsc::UnboundedSender;
